Make extendClaim public and move lastClaimExtension update
Make extendClaim public and move lastClaimExtension update. The former
will ensure the Coordinator is capable of invoking this task, will the
latter ensures the WorkPackage keeps control over when _it_ decides to
extend the claim or store the token for filtered events. Especially the
latter requires the lastClaimExtension to reside outside the extendClaim
 operation.

#enhancement/claim-extension-by-psep-coordinator